Have to CalledAs return the name the function was called as.
Currently CalledAs returns the name the command was called only
if it was executed, but parent command don't get executed. This
change allows the child to know how the parent was called as well.

Added new function Called() will let's the user know if the cmd was
called or not.

Updated tests accordingly.