Add config to prevent addition of copyright notice
Some projects do not add copyright headers to the tops of their files.
Right now you can specify `license: none` to prevent a license from
being generated for the file. But you will still get a line like:
  `// Copyright © 2017 NAME HERE <EMAIL ADDRESS>`

This change allows passing `author: none` to prevent that line from
being generated. The change does not alter the defaults. However, it
does slightly change the formatting templates to prevent the
extraneous addition of newlines when an author or liciece line is not
requested.