Add DisableHelpSubCommand option
This option stops the creation of a "help" subcommand for when a CLI
tool only wants to have "-h" and "--help" to get examples/usage.

This can (in certain situations) help reduce abiguity in documentation.
Instead of having two ways to get help about a subcommand ("app help foo"
and "app foo --help") there can be only one.

As this option defaults to false, the default behavior remains the same.