Adjust bash completion instructions for compatibility with bash 3.2
Bash 3.2 is the default shell on macOS Mojave and other OSs, but the old instructions for sourcing completion scripts into one's shell don't work with it:

```
. <(cmd completion)
```

This is because the `source` (`.`) builtin doesn't work with process substitution in 3.2. https://lists.gnu.org/archive/html/bug-bash/2006-01/msg00018.html

The `eval` approach is portable between different bash versions.