Complete hidden commands flags and its subcommands
This change enable to have a correct completion on hidden commands flags and
its subcommands if present, when using bash.
If an exact hidden command is typed on the command line and then the
user request completion, instead of printing other available subcommand
at the same level (which is thus incorrect), it will print any of its
subcommand if available.
Flags completion request (starting with -) will also work as expected,
as well as flag completion request on any subcommand (persistent flags
or local).