Uppercase first char in short desc of generate command
This commit provides consistency of the letter casing in the short
description between commands. Both the short description as well as
documented examples have the first character in the short description in
uppercase. The generate command was the exception by having the first
character of the short description in lowercase, creating an
inconsistent help menu for the user:

```sh
Available Commands:
  add         Add a command to a Cobra Application
  completion  generate the autocompletion script for the specified shell
  help        Help about any command
  init        Initialize a Cobra Application
```

* User Guide Example: https://github.com/spf13/cobra/blob/v1.2.1/user_guide.md#example
* Help command: https://github.com/spf13/cobra/blob/master/command.go#L1080