command: move required flag check before PreRun(E)
The current implementation checks the presence of required flags after
PreRun or PreRunE functions are run. This commit moves that validation
to happen before those steps.

There are use cases where we want to use flags inside PreRun/PreRunE
functions. Checking for presence of required flags in those cases
involves manually iterating and checking the flags inside your prerun
functions. If we move the validation up, we get the validation for free.

The only "counter" usecase would be when users use PreRun/PreRunE to set
required flags. This, however seems like a poor usecase that defeats the
purpose of having required flags and/or default values in the first
place.