fix: quote args in fish shell completion (#2285)
* fix: quote args in fish shell completion

The arguments used for dynamic completions in fish were missing quotes.
This resulted in an error when one of the arguments had a `*` in it
whose expansion matched no file. Once the completion got triggered, fish
errored with the message `No matches for wildcard [...]`. Don't quote
the first argument as that is the name of the binary.

* Protect against arguments that start with --

Signed-off-by: Marc Khouzam <marc.khouzam@gmail.com>

---------

Co-authored-by: Marc Khouzam <marc.khouzam@gmail.com>